Fixed warnings converting from uint32 to size_t and vice-versa Fixed some ">" and "<" testing int32 on uint32

+++ src/eve-common/cache/CachedObjectMgr.cpp
@@ -361,7 +361,7 @@ bool CachedObjectMgr::SaveCachedToFile(const std::string &cacheDir, const PyRep 
     header.timestamp = res->second->timestamp;
     header.version = res->second->version;
     header.magic = CacheFileMagic;
-    header.length = res->second->cache->content().size();
+    header.length = (uint32)res->second->cache->content().size();
 
     if(fwrite(&header, sizeof(header), 1, f) != 1) {
         fclose(f);
