+++ src/eve-core/utils/utils_hex.h
@@ -40,9 +40,9 @@ extern const uint32 HEX_DUMP_PREVIEW_LIMIT;
  */
 void build_hex_line( const uint8* buffer, size_t length, size_t offset, char* ret, unsigned int padding = 4 );
 
-void pfxHexDump( const char* pfx, FILE* into, const uint8* data, uint32 length );
-void pfxHexDump( const char* pfx, LogType type, const uint8* data, uint32 length );
-void pfxHexDumpPreview( const char* pfx, FILE* into, const uint8* data, uint32 length );
-void pfxHexDumpPreview( const char* pfx, LogType type, const uint8* data, uint32 length );
+void pfxHexDump( const char* pfx, FILE* into, const uint8* data, size_t length );
+void pfxHexDump( const char* pfx, LogType type, const uint8* data, size_t length );
+void pfxHexDumpPreview( const char* pfx, FILE* into, const uint8* data, size_t length );
+void pfxHexDumpPreview( const char* pfx, LogType type, const uint8* data, size_t length );
 
 #endif // UTILS_HEX_H
