+++ src/eve-server/inventory/EVEAttributeMgr.cpp
@@ -524,8 +524,8 @@ bool AttributeMap::ResetAttribute(uint32 attrID, bool notify)
     EvilNumber attrVal;
     uint32 attributeID;
 
-    int amount = res.GetRowCount();
-    for (int i = 0; i < amount; i++)
+    size_t amount = res.GetRowCount();
+    for (size_t i = 0; i < amount; i++)
     {
         res.GetRow(row);
         attributeID = row.GetUInt(1);
@@ -576,8 +576,8 @@ bool AttributeMap::Load()
 
     DBResultRow row;
 
-    int amount = res.GetRowCount();
-    for (int i = 0; i < amount; i++)
+    size_t amount = res.GetRowCount();
+    for (size_t i = 0; i < amount; i++)
     {
         EvilNumber attr_value;
         res.GetRow(row);
