+++ src/eve-server/inventory/ItemFactory.cpp
@@ -42,7 +42,7 @@ ItemFactory::~ItemFactory() {
     // items
     {
         std::map<uint32, InventoryItemRef>::const_iterator cur, end;
-		uint32 total_item_count = m_items.size();
+		uint32 total_item_count = (uint32)m_items.size();
 		uint32 items_saved = 0;
 		float current_percent_items_saved = 0.0;
         cur = m_items.begin();
