+++ src/eve-server/ship/TargetManager.h
@@ -53,7 +53,7 @@ public:
     SystemEntity *GetFirstTarget(bool need_locked);
     bool HasNoTargets() const { return(m_targets.empty()); }
     bool IsTargetedBySomething() const { return(!m_targetedBy.empty()); }
-    uint32 GetTotalTargets() const { return m_targets.size(); }
+    uint32 GetTotalTargets() const { return (uint32)m_targets.size(); }
 
     SystemEntity *GetTarget(uint32 targetID, bool need_locked=true) const;
     void QueueTBDestinyEvent(PyTuple **up) const;    //queue a destiny event to all people targeting me.
